correcting award table

+++ src/main/java/com/example/rewards/model/Item.java
@@ -11,9 +11,7 @@ import lombok.Setter;
 public class Item {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private Integer id;
     private String name;
-
     private String description;
 }
